requirements.txt
feedparser
newspaper3k
pydantic
langchain
langgraph
sqlalchemy
python-dotenv
streamlit
OpenAI
langchain_openai
langchain-community
selenium
html5lib